refactored main.rb file
